+++ examples/qft.rs
@@ -15,47 +15,51 @@
 
 use quantr::{
     states::{ProductState, SuperPosition},
-    Circuit, Measurement, Printer, QuantrError, Gate,
+    Circuit, Gate, Measurement, Printer, QuantrError,
 };
 
 fn main() -> Result<(), QuantrError> {
-    let mut qc: Circuit = Circuit::new(5)?;
+    let mut qc: Circuit = Circuit::new(3)?;
 
     // Apply qft
-    qc.add_repeating_gate(Gate::H, &[0, 1, 2])?
-        .add_gate(Gate::Custom(qft, &[0, 1], "QFT".to_string()), 2)? // QFT on bits 0, 1 and 2
-        .add_gate(Gate::CNot(1), 3)?
-        .add_gate(Gate::CNot(2), 4)?;
+    qc.add_repeating_gate(Gate::X, &[1, 2])?
+        .add_gate(Gate::Custom(qft, &[0, 1], "QFT".to_string()), 2)?; // QFT on bits 0, 1 and 2
 
     let mut printer = Printer::new(&qc);
     printer.print_diagram();
 
+    qc.toggle_simulation_progress();
+
     qc.simulate();
 
-    if let Measurement::Observable(bin_count) = qc.repeat_measurement(100).unwrap() {
-        for (state, count) in bin_count {
-            println!("|{}> : {}", state.to_string(), count);
+    if let Measurement::NonObservable(final_sup) = qc.get_superposition().unwrap() {
+        println!("\nThe final superposition is:");
+        for (state, amplitude) in final_sup.into_iter() {
+            println!("|{}> : {}", state.to_string(), amplitude);
         }
     }
 
     Ok(())
 }
 
-// A QFT implementation that can be used for other circuits.
+// A QFT implementation that can be used for other circuits. Note, the output is reveresed, swap
+// gates are needed.
 fn qft(input_state: ProductState) -> SuperPosition {
     let qubit_num = input_state.qubits.len();
     let mut mini_circuit: Circuit = Circuit::new(qubit_num).unwrap();
 
     for pos in 0..qubit_num {
         mini_circuit.add_gate(Gate::H, pos).unwrap();
-        for k in 1..(qubit_num - pos) {
+        for k in 2..=(qubit_num - pos) {
             mini_circuit
-                .add_gate(Gate::CRk(k as i32, k), pos)
+                .add_gate(Gate::CRk(k as i32, pos + k - 1), pos)
                 .unwrap();
         }
     }
 
-    mini_circuit.simulate();
+    mini_circuit
+        .simulate_with_register(input_state.into_super_position())
+        .unwrap();
 
     if let Measurement::NonObservable(super_pos) = mini_circuit.get_superposition().unwrap() {
         super_pos.clone()